+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-07-14 09-04-58 ]] --
+-- [[ Exported at 2022-07-15 22-48-22 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --